# Reduce ThicknessConverter allocs to minimum and improve conversion performance