## Description

Simplifies code in `CanConvertTo`/`ConvertFrom`/`ConvertTo`, removes additional type casts by using `is` checks/pattern match. Commits should be easy to review.

Uses `DefaultInterpolatedStringHandler` instead of `StringBuilder` to handle conversion in `ToString` with a pre-allocated on-stack buffer (not using interpolated strings directly due to the `NaN`->`Auto` conversion which would box as Roslyn won't handle that one.

Also stackallocs 4 doubles instead of allocating an array in `FromString`, this uses unsafe quotation due to how the current code-gen with Span<double> looks like; additional perf improvement. I didn't include benchmark/tests in this one, they're together in the other PR for `LengthConverter` #9364.

### Multiple decimals point (edge case but within original range)

```c#
ToString(new Thickness(3.33371, 2.25888, 6.6666, 8.444), CultureInfo.InvariantCulture);
```

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR_EDIT |  348.2 ns |    2.56 ns |     2.27 ns | 0.0048 |          80 B |
| Original |  393.2 ns |    3.05 ns |     2.70 ns | 0.0257 |         432 B |

### Single decimal point, standard use-case

```c#
ToString(new Thickness(1, 2, 2.5, 4), CultureInfo.InvariantCulture);
```

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR_EDIT |  277.4 ns |    2.22 ns |     2.08 ns | 0.0024 |          40 B |
| Original |  328.1 ns |    4.37 ns |     4.09 ns | 0.0205 |         344 B |

### Conversion from `double.NaN` to `Auto`

```c#
ToString(new Thickness(double.NaN), CultureInfo.InvariantCulture);
```

| Method     | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|----------- |----------:|-----------:|------------:|-------:|--------------:|
| PR_EDIT |  24.23 ns |   0.187 ns |    0.146 ns | 0.0038 |          64 B |
| Original |  32.90 ns |   0.441 ns |    0.391 ns | 0.0157 |         264 B |

## Customer Impact

Improved performance, greatly decreased allocations.

## Regression

No.

## Testing

Local build, basic set of assert testing.

```c#
CultureInfo CachedNorwegian = new CultureInfo("nn-NO");

Assert.AreEqual(ToString(new Thickness(1.5, 2, 1.25, 3), CultureInfo.InvariantCulture), ToString2(new Thickness(1.5, 2, 1.25, 3), CultureInfo.InvariantCulture));
Assert.AreEqual(ToString(new Thickness(3, 2, 6, 8), CultureInfo.InvariantCulture), ToString2(new Thickness(3, 2, 6, 8), CultureInfo.InvariantCulture));
Assert.AreEqual(ToString(new Thickness(3.33371, 2.25888, 6.6666, 8.444), CultureInfo.InvariantCulture), ToString2(new Thickness(3.33371, 2.25888, 6.6666, 8.444), CultureInfo.InvariantCulture));

//Culture with decimal comma separator
Assert.AreEqual(ToString(new Thickness(1.5, 2, 1.25, 3), CachedNorwegian), ToString2(new Thickness(1.5, 2, 1.25, 3), CachedNorwegian));
Assert.AreEqual(ToString(new Thickness(3, 2, 6, 8), CachedNorwegian), ToString2(new Thickness(3, 2, 6, 8), CachedNorwegian));
Assert.AreEqual(ToString(new Thickness(3.33371, 2.25888, 6.6666, 8.444), CachedNorwegian), ToString2(new Thickness(3.33371, 2.25888, 6.6666, 8.444), CachedNorwegian));

Assert.AreEqual(ToString(new Thickness(), CachedNorwegian), ToString2(new Thickness(), CachedNorwegian));

Assert.AreEqual(ToString(new Thickness(double.NaN), CachedNorwegian), ToString2(new Thickness(double.NaN), CachedNorwegian));
```

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9363)